# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Yap
  version: '1.0'
  contact:
    name: Marco Tormento
    email: shanefalco80@gmail.com
  description: Yet Another Pokedex
servers:
  - url: 'http://localhost:8080'
paths:
  '/pokemon/{name}':
    get:
      summary: Pokemon Information
      tags:
        - standard
      responses:
        '200':
          description: Translated Pokemon Information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      description: Returns standard Pokemon description and additional information
      operationId: getPokemonInfo
    parameters:
      - schema:
          type: string
        name: name
        in: path
        required: true
        description: Pokemon Name
  '/pokemon/translated/{name}':
    parameters:
      - schema:
          type: string
        name: name
        in: path
        required: true
        description: Pokemon name
    get:
      summary: Translated Pokemon Information
      tags:
        - translated
      responses:
        '200':
          description: Translated Pokemon Information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      description: Return translated Pokemon description and other basic information
      operationId: getTranslatedPokemonInfo
components:
  schemas:
    Pokemon:
      description: Pokemon Information
      type: object
      x-examples: {}
      properties:
        name:
          type: string
        description:
          type: string
        habitat:
          type: string
        isLegendary:
          type: boolean
    ErrorResponse:
      description: ''
      type: object
      x-examples:
        example-1:
          timestamp: '2021-06-27T16:42:15.600+0000'
          path: /pokemon/translated/chari
          error: pokemon 'chari' does not exist
          requestId: c1a8ee28-1
      properties:
        message:
          type: string
          minLength: 1
      required:
        - message